Refactor article handling and UI components
Removed `AddArticlesStockCommand`, `AddArticlesStocksCommandHandler`, `ArticleChild`, `GetArticleChildrenQuery`, and `GetArticleChildrenQueryHandler`. Renamed `AddArticleField` to `AddArticle` in `AddArticlesBox` and added methods for handling article children and clearing fields. Changed `quantity` type in `ArticleField` to `decimal`. Removed `ArticleModal.Designer.cs`.

Refactored `ArticleModal.cs` to remove old article handling code. Replaced `ExecuteGetAllQuery` with `ExecuteGetEntitiesQuery` in various modal files. Updated `EntityEditionModal.cs` methods to return `ErrorOr` types and renamed `ExecuteGetAllQuery` to `ExecuteGetEntitiesQuery`.

Added new command and handler files for article stock and children management. Introduced `ErrorOr` type for error handling. Added new `ArticleModal` class with comprehensive UI and backend logic for managing articles, including various attributes and validation.

+++ LogicPOS.Api/Features/Articles/Articles/Compositions/GetArticleChildren/ArticleChild.cs
@@ -5,6 +5,6 @@ namespace LogicPOS.Api.Features.Articles.GetArticleChildren
     public class ArticleChild
     {
         public Article Article { get; set; }
-        public uint Quantity { get; set; }
+        public decimal Quantity { get; set; }
     }
 }
